requirements.txt
aenum==3.1.11
appdirs==1.4.4
APScheduler==3.9.1
arrow==1.2.3
asgiref==3.5.2
aspose-words==23.7.0
async-generator==1.10
attrs==21.4.0
azure-core==1.25.1
azure-storage-blob==12.13.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
blinker==1.6.2
Bootstrap-Flask==2.3.0
bs4==0.0.1
case-conversion==2.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.0.12
click==8.1.7
clipboard==0.0.4
colorama==0.4.5
colour==0.1.5
cryptography==37.0.2
cursor==1.3.5
cycler==0.11.0
decorator==4.4.2
dill==0.3.5.1
Django==4.1.1
et-xmlfile==1.1.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Dropzone==1.6.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.29.1
fuzzywuzzy==0.18.0
gender-guesser==0.4.0
gevent==23.7.0
greenlet==2.0.2
h11==0.13.0
html-text==0.5.2
idna==3.3
imageio==2.22.0
imageio-ffmpeg==0.4.7
infi.systray==0.1.12
inflect==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.3.2
lxml==4.7.1
MarkupSafe==2.1.3
matplotlib==3.5.1
maybe-else==0.2.1
mbstrdecoder==1.1.1
MetaTrader5==5.0.45
MouseInfo==0.1.3
moviepy==1.0.3
msal==1.19.0
msoffcrypto-tool==5.0.0
msrest==0.7.1
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.0
ntlm-auth==1.5.0
numpy==1.24.1
O365==2.0.20
oauthlib==3.2.1
office365==0.3.15
Office365-REST-Python-Client==2.3.14
olefile==0.46
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pandas-datareader==0.10.0
parsedatetime==2.6
pathmagic==0.3.14
patsy==0.5.2
Pillow==9.0.1
plotly==5.10.0
plotly-express==0.4.1
prettierfier==1.0.3
prettytable==3.3.0
proglog==0.1.10
protobuf==4.21.12
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.10.2
pydub==0.25.1
PyGetWindow==0.0.9
pyinstrument==4.3.0
pyiotools==0.3.18
PyJWT==2.5.0
pymiscutils==0.3.14
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pypdf==3.14.0
PyPDF2==2.10.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pysubtypes==0.3.18
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==304
readchar==4.0.3
regex==2022.9.13
requests==2.27.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.12.0
selenium==4.2.0
Send2Trash==1.8.0
SharePlum==0.5.1
simplejson==3.17.6
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.20
sqlparse==0.4.2
statsmodels==0.13.2
stringcase==1.2.0
tabulate==0.8.10
tenacity==8.0.1
threadpoolctl==3.1.0
tika==2.6.0
tqdm==4.64.1
trio==0.20.0
trio-websocket==0.9.2
typepy==1.3.0
types-cryptography==3.3.23
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
Unidecode==1.3.6
urllib3==1.26.8
validate-docbr==1.10.0
wcwidth==0.2.5
webdriver-manager==3.8.3
Werkzeug==2.3.7
wsproto==1.1.0
WTForms==3.0.1
XlsxWriter==3.0.3
zope.event==5.0
zope.interface==6.0
